refactor(api): move notification-preferences to modules/ as controller, model, service
